fix: Change typeVersion to double

+++ packages/@n8n/db/src/migrations/common/1770000000000-CreateChatHubToolsTable.ts
@@ -47,7 +47,7 @@ export class CreateChatHubToolsTable1770000000000 implements ReversibleMigration
 				column('id').uuid.primary,
 				column('name').varchar(255).notNull,
 				column('type').varchar(255).notNull,
-				column('typeVersion').int.notNull,
+				column('typeVersion').double.notNull,
 				column('ownerId').uuid.notNull,
 				column('definition').json.notNull,
 				column('enabled').bool.notNull.default(true),
